requirements.txt
python-dotenv
Flask
Flask-Limiter
imap-tools
requests